Rename conch_blastlist_new -> conch_blastlist_from_result_set
In order to be able to test things that depend on blastlists without
also having to create result_sets, we want to add a constructor,
conch_blastlist_new, that doesn't require a resultset.

This rename moves the extant conch_blastlist_new out of the way so we
can do that.