Rename conch_listview_new_from_args
With the previous name, this looks like an alternative constructor, which should really be in the listview file. However, it's not really a constructor, it's a method that parses command line arguments and then creates a new listview using the constructor. In order to clarify the behaviour here, we've renamed the method, and had it return a struct of the parsed cli options, and then constructed a new listview in the normal way.

At the moment there is only one command line option but this format allows for more to be added in future (though the constructor would have to change).