Notifications backend API
The idea is that instead of polling the database to find out when
there is new data, we have a notification value which we can
receive into that will tell us if there have been any inserts into
the database since we last asked. If there have been, we update
then.

This gives us immediate updates instead of having to wait 10
seconds.