Add the capability to return blast lines
Wordwrap returns the words one by one and their positions on the screen. This new function returns the lines that a blast is split up into according to the available width, that you pass in.

At the moment it is just the blast content. Possibly it may include attachments and the attribution line.

This is not yet used anywhere.

We are not sure if we need to free the pointer declared on line 15, and if so, where...