Refactor line-wrap to return array of lines
Previously, the caller was responsible for creating the array in which the lines made their homes, and the line-wrap function edited this in place. Now the function returns a new array; in a later commit, we'll create a function to free this array to deal with the memory leak (which already existed in the prior version).