Refactor line-wrap to avoid pointing to a constant
While attempting to write a `free` function, it emerged that final lines handled by this function weren't being copied to our memory space but, instead, we were pointing to constants owned by the caller. This refactor addresses that; unfortunately it breaks if the blast is empty. We'll fix this in a later commit.