Move line-wrap function to `strutils`
This is actually a more generic function that takes a string and an available width and returns an array of lines that the string splits into. This commit just moves to its natural home; in a later commit we will rename it, but wanted to keep this commit clear.

Interestingly, we added so many tests to `strutils` that git thought we were renaming the file. Go us!