Rename line wrap function to `wrap_lines`
This is a more generic function than the previous name implied. This commit changes the method signature, i.e. the name and the parameters, and the test names.

I haven't changed the local variables in the tests, as this is the case we are actually interested in testing, but will do so if someone objects.

Also added the width as a variable to make the tests clearer.