Change return value for line wrapping method
This is part of a longer change; we're going to change the name of the method, and pull out the render part so callers will have to call both, but I'm tired, so we'll do that tomorrow. Getting this change in now in case it affects what others are doing. No-one else is using the return type yet.