Add a rudimentary Vagrantfile
This file uses a shell provisioner to install all the packages you need
to run conch. There are a few manual steps at the end in order to be
able to run the tests; these could be handled by the provisioner but
it would make more sense to use puppet or chef for this. I may implement
them in shell later for a laugh.

There is an option to use a basebox that exists on core.fort so this
can be built afresh while still here.

You still have to set up your gitconfig and vimrc etc; it does, at least, installvim :)