An example post-receive hook
If you want to see git commits on your publicly viewable conch then place this file in hooks/post-receive

It will post a blast for the newest commit to the master branch, posted as the user who made the commit. (you will need to set up a .mailmap file to match up the usernames)

The blast will include the abbreviated sha of the commit and the first line of the commit.

It will also include the `git log` output in the extended section of the blast.