Better navigation: Arrows and paging.
Support up/down for j/k. More interestingly support page up and
page down.

Paging is made a bit weird by the fact that we can't reliably
consider the page size in the view. Instead we have a dynamically
adapting page size that uses the info we have at the start of the
call when the render loop has nicely set everything up for us.

This is mostly with a goal of making sure you never miss posts when
paging, with a secondary goal that wherever possible paging up then
down or vice versa should be a no-op.