Add strcopytrunc for making a truncated copy of a string
We're about to use this to ensure that attachment URLs and attribution
lines don't overflow the edge of the screen on narrow terminals.

Extracted into a "library function" because strncpy(3) doesn't add a NUL
to the end of the copied string when the source string is longer than
the destination (i.e. when truncating). This makes it reasonably easy to
create an unterminated string when trying to truncate.