Switch render direction only when absolutely necessary
Previously, if we moved the cursor from (say) the penultimate blast to
the last blast on screen, the render direction would be changed
unconditionally.

Now we check if the last blast was fully rendered. If it was, there's no
need to change render direction until we hit "j" again, and we select a
blast that is currently off screen.