Mock fatal_error in tests.
We have essentially implemented `assert_raises` in C.

When a piece of code needs to call `fatal_error`, in the check binary instead of linking to `explode.o` link to `check-explode.o`.

There is a `jmp_buf` defined in `checkrunner.h` that you can use to store a `setjmp` location in your tests.

Eg.

  bool caught = false;
  if (setjmp(check_jump) == 0) {
    // code that will cause fatal_error to be called.
  } else {
    caught = true;
  }
  ck_assert_msg(caught, "fatal error was not called");