Stop extended marker being exponentially wrong
When we figure out where the extended marker is for colouring it in, we were previously setting a variable called `marker_rel_y` to indicate it's y position relative to the top of the blast, i.e. what line it is on. However, when rendering, we were then adding this to y, so as the message wrapped it got more and more incorrect: at line 0 it was correct, if on line one it was 1 out, if on line two, 2 out, etc.

In fact, we don't want to dictate the y position, we just want to know what line the last line of the blast is on so we can determine in `conch_blast_render` when we are on the right line.

This commit fixes that bug, and also renames the variable to `content_last_line` as that is what we really care about.