Automate remaining steps in Vagrantfile
Just copy the `pg_hba.conf` we require, rather than messing around with sed.

Moved the remaining manual steps to the README, as they are fine-tuning you
can do to the VM rather than steps to be automated.

This now brings up a machine that you can run conch and `make check` on.