Make wrap_lines newline-aware
Previously, wrap_lines would not treat newline characters specially:
they would either be included as literal values, or removed when at a
wrap boundary.

This commit substantially reworks wrap_lines, both to fix this
limitation (newlines are now preserved, but not as literals, and the
count in `nlines` is correctly updated), but also to remove magic
constants and unnecessary block allocations.