Increase timeout for one test
This test regularly times out when run from a VM. The default timeout is 4 seconds
and 6 seconds is enough, so increase the timeout on just this test.

This means I can't use the macro so have to define the test separately.

I've also removed the line from the README because it was just this one test that
was timing out and only for me.