Interpret options in order of (defaults, config file, command line options)
I’ve tried to simplify* the determining of configuration values.

For each command there is a function to create an empty config struct (because clang won’t let me use the empty initializer {}).
Then there is a function that will set the built in default values.

When setting up the options:
  1. create an options object to store final results
  2. create an options object to store the default results
  3. call defaulting function on default options
  4. call zeroing function on final options
  5. parse command line arguments
  6. use command line arguments and defaults to determine the configuration file location
  7. parse the configuration file.
  8. merge each option with the precedence:
    command line || config file || default

*via the medium of adding a lot more code.