Delete white marker as it doesn't parse unicode
`strlen` doesn't interpret unicode characters as one would expect
so when there is a unicode character the white extended marker
displays in one place but reports an incorrect string length so the
magenta marker doesn't overwrite it in the same place.

This commit deletes the white marker after measuring it, and adds
a test to document this behaviour.

It does mean that the magenta marker will be slightly in the wrong place
but at least the white one won't be showing.