Rework word-wrap
`23d1c56` completely replaced the code but unfortunately there was no
explanation of why or what improvements had been made. It's not clearer or
more readable in the new version, though it may well be more efficient.

Since the purpose of the fort is both to produce a cool thing but also and
more importantly to learn and improve our C, I'm reimplementing the previous
code with the aim of learning how to improve it, and my C.

This commit reinstates the code as was, so does not include the two features
subsquently implemented: newlines awareness, and not stripping leading spaces.
I've commented out the relevant tests, and will reimplement these features in
subsequent commits.

Future changes and suggestions to this file very welcome of course, but please
accompany them with explanations! Especially when off-fort. Smaller commits
preferred so working can be followed.