Special case: respect spaces after newline wrap
I'm not convinced this behaviour makes sense; we only respect spaces in a line
if they immediately follow a new line, but other spaces are almost arbitrarily
cleared or not, depending on where they fall relative to the line break.

However, this behaviour appears to be required for the help screen display, so
I've reimplemented it as it was, but I suspect there should probably be some
refactoring of the help display.