Change name of test to make intent clearer
My first implementation stopped stripping leading spaces, and several
other tests failed too. The intention here is clearly that if a new
line starts with a space or spaces, they should not be stripped, but
spaces at end of lines should be. This commit renames the test to make
the intent clearer; it is still commented out though so failing tests are
not committed to master.